Segundo intento funcional modulado
Se realiza la modularización del esquema, modelo y petición.
