+++ server/src/controllers/pruebaControllers.js
@@ -1,15 +1,4 @@
-const mongoose = require('mongoose')
-const Schema = mongoose.Schema
-
-const student = new Schema({
-  nombre: String,
-  carnet: Number,
-  planAcademico: String
-},
-{
-  collection: 'Estudiante'
-})
-const studentModel = mongoose.model('studentModel', student)
+const studentModel = require('../models/Estudiante')
 module.exports = {
   async get (req, res) {
     try {
