+++ server/src/models/Estudiante.js
@@ -1,5 +1,13 @@
-// const mongoose = require('mongoose')
-// const Schema = mongoose.Schema
-module.exports = (Schema) => {
+const mongoose = require('mongoose')
+const Schema = mongoose.Schema
 
-}
+const student = new Schema({
+  nombre: String,
+  carnet: Number,
+  planAcademico: String
+},
+{
+  collection: 'Estudiante'
+})
+const studentModel = mongoose.model('studentModel', student)
+module.exports = studentModel
